+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto.java
@@ -33,24 +33,4 @@ public class Auto extends ULLinearOpMode {
         // tell the robot to map out the path and follow it
         robot.followCurve(allPoints, Math.toRadians(90));
     }
-
-    @TeleOp(name = "GamepadTester", group = "OpModes")
-    public static class GamepadTester extends ULOpMode {
-
-        Robot robot;
-
-        @Override
-        public void init() {
-            robot = new Robot();
-        }
-
-        @Override
-        public void loop() {
-            telemetry.addData("Gamepad1 Left Joystick Y:   ", gamepad1.left_stick_y);
-            telemetry.addData("Gamepad1 Left Joystick X:   ", gamepad1.left_stick_x);
-            telemetry.addData("Gamepad1 Right Joystick Y:   ", gamepad1.right_stick_y);
-            telemetry.addData("Gamepad1 Right Joystick X:   ", gamepad1.right_stick_x);
-            telemetry.update();
-        }
-    }
 }
